refactor(tests): organize/merge existing equipment route tests
