+++ tests/fixtures/equipment/addEquimentForm.ts
@@ -5,3 +5,4 @@ export const newEquipmentFixture = {
     availability: 'true',
     workshopPrerequisites: ''
 }
+
