+++ tests/fixtures/equipment/getEquipmentSlotsWithStatus.ts
@@ -18,3 +18,4 @@ export const getEquipmentSlotsWithStatusFixture = [
      }
    },
 ]
+
