+++ tests/routes/dashboard/addequipment.test.ts
@@ -1,6 +1,6 @@
 import { getWorkshopsFixture } from 'tests/fixtures/workshop/workshops';
 import { getRoleUserFixture, getRoleUserAdminFixture } from 'tests/fixtures/session/getRoleUser';
-import { newEquipmentFixture } from 'tests/fixtures/equipments/addEquimentForm';
+import { newEquipmentFixture } from 'tests/fixtures/equipment/addEquimentForm';
 import { getEquipmentByName, addEquipment } from '~/models/equipment.server';
 import { loader, action } from '~/routes/dashboard/addequipment';
 import { getRoleUser } from "~/utils/session.server";
